Add sleep mode.  Move startup/shutdown into functions.  Add tube LED feature.
Unit starts up sleeping and either a knob rotation (interrupt in the Encoder library) or a knob press will wake it up.  Had to move pins around so I could use an interrupt pin for the button.  This didn't cause a noticeable difference in the encoder performance.  Unit now has functions for startup and shutdown animation routines.